moved air extras to common code
