+++ ens/modules/airstack/common/index.ts
@@ -8,6 +8,7 @@ import {
   AirEntityCounter,
   AirMeta,
   AirAccount,
+  AirExtra,
 } from "../../../generated/schema";
 
 export const AIR_META_ID = "AIR_META";
@@ -174,4 +175,26 @@ export function getOrCreateAirAccount(chainId: string, address: string, block: A
     entity.createdAt = block.id;
   }
   return entity as AirAccount;
+}
+
+/**
+ * @dev this function does not save the returned entity
+ * @dev this function creates an air extra entity
+ * @param name air extra name
+ * @param value air extra value
+ * @param extraId air extra entity id
+ * @returns air extra entity
+ */
+export function createAirExtra(
+  name: string,
+  value: string,
+  id: string,
+): AirExtra {
+  let entity = AirExtra.load(id);
+  if (entity == null) {
+    entity = new AirExtra(id);
+    entity.name = name;
+    entity.value = value;
+  }
+  return entity as AirExtra;
 }
